[TSK-84] Feat: 유저 활동 지역 추가

+++ src/main/kotlin/com/wafflestudio/team2server/area/service/AreaServiceImpl.kt
@@ -10,13 +10,13 @@ class AreaServiceImpl(
 	private val areaRepository: AreaRepository,
 	private val areaAdjRepository: AreaAdjRepository
 ) : AreaService {
-	override fun getAdjAreas(id: Long, distance: Int): Set<Long> {
+	override fun getAdjAreas(id: Int, distance: Int): Set<Int> {
 		return areaAdjRepository.getAreaAdjEntityByAreaIdAndDistanceIsLessThanEqual(id, distance)
 			.map { it -> it.targetId }
 			.toSet()
 	}
 
-	override fun getAreaById(id: Long): AreaEntity {
+	override fun getAreaById(id: Int): AreaEntity {
 		return areaRepository.findById(id).get()
 	}
 }
